Rename module with yaml reader
