+++ acid/settings_reader.py
